+++ pyannote/audio/models/embedding/wespeaker/__init__.py
@@ -95,6 +95,11 @@ class BaseWeSpeakerResNet(Model):
 
         return features - torch.mean(features, dim=1, keepdim=True)
 
+    @property
+    def dimension(self) -> int:
+        """Dimension of output"""
+        return self.resnet.embed_dim
+
     def forward(
         self, waveforms: torch.Tensor, weights: Optional[torch.Tensor] = None
     ) -> torch.Tensor: