+++ pyannote/audio/models/segmentation/PyanNet.py
@@ -20,7 +20,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-
+from functools import cached_property, lru_cache
 from typing import Optional
 
 import torch
@@ -139,6 +139,17 @@ class PyanNet(Model):
             ]
         )
 
+    @property
+    def dimension(self) -> int:
+        """Dimension of output"""
+        if isinstance(self.specifications, tuple):
+            raise ValueError("PyanNet does not support multi-tasking.")
+
+        if self.specifications.powerset:
+            return self.specifications.num_powerset_classes
+        else:
+            return len(self.specifications.classes)
+
     def build(self):
         if self.hparams.linear["num_layers"] > 0:
             in_features = self.hparams.linear["hidden_size"]
@@ -147,17 +158,10 @@ class PyanNet(Model):
                 2 if self.hparams.lstm["bidirectional"] else 1
             )
 
-        if isinstance(self.specifications, tuple):
-            raise ValueError("PyanNet does not support multi-tasking.")
-
-        if self.specifications.powerset:
-            out_features = self.specifications.num_powerset_classes
-        else:
-            out_features = len(self.specifications.classes)
-
-        self.classifier = nn.Linear(in_features, out_features)
+        self.classifier = nn.Linear(in_features, self.dimension)
         self.activation = self.default_activation()
 
+    @lru_cache
     def num_frames(self, num_samples: int) -> int:
         """Compute number of output frames for a given number of input samples
 
@@ -174,6 +178,7 @@ class PyanNet(Model):
 
         return self.sincnet.num_frames(num_samples)
 
+    @cached_property
     def receptive_field(self) -> SlidingWindow:
         """Compute receptive field
 
@@ -186,7 +191,7 @@ class PyanNet(Model):
         https://distill.pub/2019/computing-receptive-fields/
 
         """
-        return self.sincnet.receptive_field()
+        return self.sincnet.receptive_field
 
     def forward(self, waveforms: torch.Tensor) -> torch.Tensor:
         """Pass forward