+++ pyannote/audio/pipelines/overlapped_speech_detection.py
@@ -128,7 +128,7 @@ class OverlappedSpeechDetection(Pipeline):
         # load model
         model = get_model(segmentation, use_auth_token=use_auth_token)
 
-        if model.example_output.dimension > 1:
+        if model.dimension > 1:
             inference_kwargs["pre_aggregation_hook"] = lambda scores: np.partition(
                 scores, -2, axis=-1
             )[:, :, -2, np.newaxis]