+++ pyannote/audio/tasks/segmentation/voice_activity_detection.py
@@ -164,18 +164,25 @@ class VoiceActivityDetection(SegmentationTask):
         ]
 
         # discretize chunk annotations at model output resolution
-        start = np.maximum(chunk_annotations["start"], chunk.start) - chunk.start
-        start_idx = np.floor(start / self.model.example_output.frames.step).astype(int)
-        end = np.minimum(chunk_annotations["end"], chunk.end) - chunk.start
-        end_idx = np.ceil(end / self.model.example_output.frames.step).astype(int)
+        step = self.model.receptive_field.step
+        half = 0.5 * self.model.receptive_field.duration
+
+        start = np.maximum(chunk_annotations["start"], chunk.start) - chunk.start - half
+        start_idx = np.maximum(0, np.round(start / step)).astype(int)
+
+        end = np.minimum(chunk_annotations["end"], chunk.end) - chunk.start - half
+        end_idx = np.round(end / step).astype(int)
 
         # frame-level targets
-        y = np.zeros((self.model.example_output.num_frames, 1), dtype=np.uint8)
+        num_frames = self.model.num_frames(
+            round(duration * self.model.hparams.sample_rate)
+        )
+        y = np.zeros((num_frames, 1), dtype=np.uint8)
         for start, end in zip(start_idx, end_idx):
-            y[start:end, 0] = 1
+            y[start : end + 1, 0] = 1
 
         sample["y"] = SlidingWindowFeature(
-            y, self.model.example_output.frames, labels=["speech"]
+            y, self.model.receptive_field, labels=["speech"]
         )
 
         metadata = self.prepared_data["audio-metadata"][file_id]